Leave invalid settings uninitialized
When we try to load a setting that fails to initialize (using the new
lazy-load function), we should leave that "uninitialized" in the UI
property grid. That means we can (a) exclude it from the settings sent
to the radio in set_settings() and (b) allow the user to give the
setting a sane value according to the limits provided in the setting
to correct the issue.